refactor: extract search methods into a new SearchService class
